photo num size

+++ frameworks/services/media_backup_extension/src/upgrade_restore.cpp
@@ -1434,7 +1434,7 @@ void UpgradeRestore::InsertFaceAnalysisData(const std::vector<FileInfo> &fileInf
             (long)(startInsertMap - startInsertFace), (long)mapRowNum, (long)(endInsert - startInsertMap));
     }
     int64_t end = MediaFileUtils::UTCTimeMilliSeconds();
-    photoNum = static_cast<uint64_t>(filesWithFace.size());
+    photoNum = filesWithFace.size();
     migratePortraitFaceNumber_ += faceRowNum;
     migratePortraitPhotoNumber_ += photoNum;
     migratePortraitTotalTimeCost_ += end - start;
